refactor: ui extensions core:common to core:ui
