+++ feature/ledger/src/main/java/com/moneymong/moneymong/ledger/LedgerScreen.kt
@@ -32,7 +32,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavOptions
 import com.example.member.MemberScreen
 import com.moneymong.moneymong.analytics.LocalAnalyticsTracker
-import com.moneymong.moneymong.common.ui.plus
+import com.moneymong.moneymong.ui.plus
 import com.moneymong.moneymong.design_system.R
 import com.moneymong.moneymong.design_system.component.bottomSheet.MDSBottomSheet
 import com.moneymong.moneymong.design_system.component.button.MDSFloatingActionButton
