+++ feature/ledgerdetail/src/main/java/com/moneymong/moneymong/ledgerdetail/LedgerDetailViewModel.kt
@@ -6,9 +6,9 @@ import com.moneymong.moneymong.common.base.BaseViewModel
 import com.moneymong.moneymong.common.error.MoneyMongError
 import com.moneymong.moneymong.common.ext.toDateFormat
 import com.moneymong.moneymong.common.ext.toMultipart
-import com.moneymong.moneymong.common.ui.isValidPaymentDate
-import com.moneymong.moneymong.common.ui.isValidPaymentTime
-import com.moneymong.moneymong.common.ui.validateValue
+import com.moneymong.moneymong.ui.isValidPaymentDate
+import com.moneymong.moneymong.ui.isValidPaymentTime
+import com.moneymong.moneymong.ui.validateValue
 import com.moneymong.moneymong.domain.usecase.ledgerdetail.FetchLedgerTransactionDetailUseCase
 import com.moneymong.moneymong.domain.usecase.ledgerdetail.DeleteLedgerDetailUseCase
 import com.moneymong.moneymong.domain.usecase.ledgerdetail.DeleteLedgerDocumentTransactionUseCase
