+++ feature/ocr-detail/src/main/java/com/moneymong/moneymong/ocr_detail/OCRDetailViewModel.kt
@@ -4,9 +4,9 @@ import android.content.SharedPreferences
 import androidx.compose.ui.text.input.TextFieldValue
 import com.moneymong.moneymong.common.base.BaseViewModel
 import com.moneymong.moneymong.common.ext.toMultipart
-import com.moneymong.moneymong.common.ui.isValidPaymentDate
-import com.moneymong.moneymong.common.ui.isValidPaymentTime
-import com.moneymong.moneymong.common.ui.validateValue
+import com.moneymong.moneymong.ui.isValidPaymentDate
+import com.moneymong.moneymong.ui.isValidPaymentTime
+import com.moneymong.moneymong.ui.validateValue
 import com.moneymong.moneymong.domain.usecase.agency.FetchAgencyIdUseCase
 import com.moneymong.moneymong.domain.usecase.ledger.PostLedgerTransactionUseCase
 import com.moneymong.moneymong.domain.usecase.ocr.PostFileUploadUseCase
